RuneRunner/requirements.txt
aiohttp==3.8.5
beautifulsoup4==4.10.0
python-dotenv==0.19.2
scikit-learn
joblib
numpy
pandas
requests
Flask
selenium==4.1.0
psycopg2-binary
sentence-transformers
jinja2